[#229] refactor: CommentHeader 컴포넌트 분리
