+++ src/app/debate-list/[debateId]/components/Comment/CommentsSection.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import { getCommentList } from '@/service/debate/comment';
 import { CommentProps } from '@/types/debate/comment';
-import ChakraSelect from '@/components/Common/ChakraSelect';
-import Comment from './CommentCard';
+import CommentsHeader from './CommentsHeader';
+import CommentCard from './CommentCard';
 
 interface CommentsSectionProps {
   debateId: number;
@@ -20,11 +20,6 @@ const CommentsSection: React.FC<CommentsSectionProps> = ({
   const [isChanged, setIsChanged] = useState<boolean>(false);
   const [selectedOption, setSelectedOption] = useState<string>('등록순');
 
-  const commentSelectOptions = [
-    { value: 'latest', label: '최신순' },
-    { value: 'oldest', label: '등록순' },
-  ];
-
   useEffect(() => {
     getCommentList(debateId)
       .then(comments => {
@@ -37,25 +32,18 @@ const CommentsSection: React.FC<CommentsSectionProps> = ({
       .catch(error => {
         console.error('Error fetching comments:', error);
       });
-  }, [debateId, commentsUpdated, isChanged]);
+  }, [debateId, commentsUpdated, isChanged, selectedOption]);
 
   return (
-    <div className="flex flex-col gap-6 mb-5  rounded-lg ">
-      <div className="flex flex-col w-full mb-10 ">
-        <div className="flex justify-between place-items-center my-3">
-          <div className="flex justify-left place-items-end">
-            <span className="text-2xl font-bold text-white">토론 현황</span>
-            <span className="text-white ml-2">(총 {commentList.length}개)</span>
-          </div>
-          <ChakraSelect
-            options={commentSelectOptions}
-            setSelectedOption={setSelectedOption}
-          />
-        </div>
-
+    <div className="flex flex-col gap-4 rounded-lg mb-16">
+      <CommentsHeader
+        commentsNum={commentList.length}
+        setSelectedOption={setSelectedOption}
+      />
+      <div className="flex flex-col">
         {commentList.map((comment: CommentProps) => {
           return (
-            <Comment
+            <CommentCard
               key={comment.commentId}
               userImg={comment.commenter.profileImgUrl}
               userName={comment.commenter.nickname}
